split report parser and message generation from main app
